requirements.txt
beautifulsoup4==4.13.4
boto3==1.38.39
botocore==1.38.39
json_repair==0.39.1
matplotlib==3.10.3
networkx==3.4.2
numpy==2.3.1
openai==1.93.0
pandas==2.3.0
pillow==11.3.0
protobuf==6.31.1
pydantic==2.11.7
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.4
seaborn==0.13.2
selenium==4.34.0
tenacity==9.1.2
tqdm==4.67.1
google-genai==1.2.0